chore(EMS-3025): moved companies house eligiblity to companies-house folder
