+++ e2e-tests/insurance/cypress/e2e/journeys/eligibility/companies-house/companies-house-number/companies-house-number-answer-no.spec.js
@@ -1,6 +1,6 @@
-import { backLink } from '../../../../../../pages/shared';
-import { LINKS } from '../../../../../../content-strings';
-import { INSURANCE_ROUTES } from '../../../../../../constants/routes/insurance';
+import { backLink } from '../../../../../../../pages/shared';
+import { LINKS } from '../../../../../../../content-strings';
+import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
 
 const {
   ELIGIBILITY: { COMPANIES_HOUSE_NUMBER, NO_COMPANIES_HOUSE_NUMBER },
