+++ e2e-tests/insurance/cypress/e2e/journeys/eligibility/companies-house/companies-house-number/companies-house-number.spec.js
@@ -1,7 +1,7 @@
-import { yesRadio, noRadio } from '../../../../../../pages/shared';
-import { PAGES, ERROR_MESSAGES } from '../../../../../../content-strings';
-import { ROUTES, FIELD_VALUES } from '../../../../../../constants';
-import { INSURANCE_FIELD_IDS } from '../../../../../../constants/field-ids/insurance';
+import { yesRadio, noRadio } from '../../../../../../../pages/shared';
+import { PAGES, ERROR_MESSAGES } from '../../../../../../../content-strings';
+import { ROUTES, FIELD_VALUES } from '../../../../../../../constants';
+import { INSURANCE_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance';
 
 const CONTENT_STRINGS = PAGES.INSURANCE.ELIGIBILITY.HAS_COMPANIES_HOUSE_NUMBER;
