+++ e2e-tests/insurance/cypress/e2e/journeys/eligibility/companies-house/companies-house-unavailable/companies-house-unavailable.spec.js
@@ -1,6 +1,6 @@
-import { companiesHouseUnavailablePage } from '../../../../../../pages/insurance/eligibility';
-import { PAGES } from '../../../../../../content-strings';
-import { INSURANCE_ROUTES } from '../../../../../../constants/routes/insurance';
+import { companiesHouseUnavailablePage } from '../../../../../../../pages/insurance/eligibility';
+import { PAGES } from '../../../../../../../content-strings';
+import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
 
 const CONTENT_STRINGS = PAGES.INSURANCE.ELIGIBILITY.COMPANIES_HOUSE_UNAVAILABLE;
